Implement a collection for to-device events
To-device events are received in a sync response in a separate field of
the response.

They are sent between dirrectly without involving a room and are not
permanently stored on the serever. They are mostly used to set up
encryption between devices.

This patch implements a enum which encapsulates many of the messages
that will appear in the to-device part of a sync response.