Properly validate localpart when parsing UserId
When a `UserId` was created via `parse_with_server_name` (and it's
variants), it was not
validating that localpart was valid. This lead to invalid user ids being
constructed and lead to a crash when `is_historical` was called.

This changes will prevent that this invalid user is can be constructed
and it also changes the `is_historical` function to prevent an `unwrap`,
defaulting to "localpart is not fully conforming" if there are invalid
characters.