macros: Apply cfg_attr for unstable-exhaustive-types feature only when it is defined on the crate
The `unexpected-cfgs` lint is triggered when we generate a
`#[cfg_attr(not(feature = "unstable-exhaustive-types"),
non_exhaustive)]` attribute and the cargo feature is not defined.
This is not a problem with the ruma crates where the feature is defined
on pretty much all crates,
but it creates problems with downstream projects that use these macros.

The solution here is to detect the presence of the cargo feature at
compile time.
If the feature is present, we generate the `cfg_attr`, otherwise we just
generate a `#[non_exhaustive]`.