client-api: Expose the ErrorCode of ErrorKind
It does not really make sense to implement `AsRef<str>` and `Display`
for `ErrorKind`
since some variants include more data than the error code.

Besides, since `ErrCode` was only used for deserialization, there was a
possibility that
the 2 types could get out of sync, resulting in variants not being able
to be deserialized.

We expose `ErrCode` as `ErrorCode` so that users that want the error
code can access it,
and we use `ErrorCode` during serialization, making sure that the types
stay in sync.