fix: Allow configure exhaustive types when environment variable is set
Allow configure exhaustive types via the
`RUMA_UNSTABLE_EXHAUSTIVE_TYPES` environment variable in the same way as
it's done in other crates.

Rumma common contains, at least, `NewPushRule` which may be configured
as exhaustive. `ruma-client-api` uses the mentioned type and, when the
environment variable is set, it fails to compile the project.