Commit Cargo.lock
With all the latest dependencies except base64ct, which has a newer
MSRV than we do.
This fixes the occassional flakes of our MSRV CI job as dependencies
bump their MSRV, at the expense of requiring us to bump the lockfile
from time to time. It also means contributors are always using the
same dependency versions.