events: relax requirement for `BundledThread::latest_event` to include a `room_id`
The spec is a bit unclear about whether the included latest event always
has a room id or not. Based on discussions with Synapse developers, it
seems the event is included as in the same context of the owning event;
if both come from sync, then they'll follow the sync format (thus won't
have a room id).

It's better to relax the format here; usually the room id should always
be available from the surrounding context anyways.