serde: Fix default_on_error
It was working with `serde_json::from_value` but not other functions
like `serde_json::from_(str/slice)`. This is due to the fact that the
serde_json deserializers parse the JSON lazily and were expecting the
end of a map after ignoring the erroneous content (due to the expected
type being a struct), instead of expecting the end matching the
beginning that was found.

To work around that, we first parse the full value as a
`Box<RawJsonValue>`, then try to deserialize the expected type from that
value. It now works will all 3 methods but is limited to deserializing
JSON.