client-api,events: Remove StrippedState and add optional fields to StrippedStateEvent instead
`StrippedState` makes it complicated to access the content of an event,
with several levels of matching (event format, event type, and for some
formats another level), to end up with up to 3 possible content types.

It is simpler for users to have a single enum for the event type and be
able to access the content and the fields directly.