api: Change Metadata struct to a trait
Its fields are now associated constants with the same name converted to
uppercase.

The `metadata!` macro generates the trait implementation for a type
named `Request` by default. The type can be changed with an `@for`
setting.

`Metadata` is a supertrait of `OutgoingRequest` and `IncomingRequest`.

The goal of this is to be able in the future to change some of the
associated constants to associated types to use them in bounds.