api: Require authentication type to be in scope
Instead of only supporting the types in ruma_common::api::auth_scheme.

This allows to use custom types. This is also necessary to complete the
implementation of `OutgoingRequest::add_authentication` of
`ServerSignatures`, we need to use the `XMatrix` type in
ruma-federation-api. One solution is to move `ServerSignatures` to that
crate.