push: Force PushConditionRoomCtx to be !RefUnwindSafe
Forcing `HasThreadSubscriptionFn` to be `RefUnwindSafe` is actually
problematic in the matrix-rust-sdk because for example
`tokio::sync::Mutex` is `!RefUnwindSafe`.

Since we cannot implement negative traits yet, this uses a `PhantomData`
to pretend like `HasThreadSubscriptionFn` is always part of the struct,
to avoid changing the auto trait implementations when the feature is
enabled.